automation/requirements.txt
asttokens==2.4.1
colorama==0.4.6
exceptiongroup==1.2.0
executing==2.0.1
icecream==2.1.3
iniconfig==2.0.0
packaging==23.2
paho-mqtt==1.6.1
pluggy==1.3.0
Pygments==2.17.2
pytest==7.4.3
six==1.16.0
tomli==2.0.1
pandas==2.2.2
PyYAML==6.0.2
python-dotenv==1.0.0
